Park doil

+++ SDL Game Framework/SDL Game Framework/Framework/Csv.h
@@ -16,7 +16,7 @@ struct Parsing_Data {
     char    SOUND_NAME[128];
 
     char    EFFECT_SOUND_NAME[128];
-    int64   EFFECT_COUNT;
+    float   EFFECT_COUNT;
     wchar_t TEXT[2000];
 
     int64   CHOOSE_1_NEXT_SCENE;
